pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "invest-system"
version = "0.1.0"
description = "Open-source quantitative investment toolset for data collection, strategy backtesting, market intelligence, and real-time monitoring"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "backtrader==1.9.78.123",
    "beautifulsoup4==4.14.3",
    "click==8.3.1",
    "curl-cffi==0.13.0",
    "flask==3.1.3",
    "numpy==2.4.3",
    "pandas==3.0.1",
    "peewee==4.0.2",
    "python-dotenv>=1.2.2",
    "requests==2.32.5",
    "telethon==1.42.0",
    "websockets==16.0",
    "yfinance==1.2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "responses>=0.26.0",
]

[project.scripts]
invest-main = "invest_system.main:run_from_cli"
invest-backtest = "invest_system.backtest:run_from_cli"
invest-webapp = "invest_system.webapp:run_from_cli"
invest-fetcher = "invest_system.data.fetcher:run_from_cli"
invest-intelligence = "invest_system.intelligence:run_from_cli"
